Move customValidationHandler
